# backend/spec/openapi.yaml
openapi: "3.1.1"
info:
  title: My Library API
  description: |
    Book library management system.
    All endpoints return JSON. Some endpoints require authentication
    (Bearer token from `/api/auth/login`).
  version: "0.3.0"
  contact:
    name: My Library
servers:
  - url: http://127.0.0.1:8000
    description: Local development
  - url: https://dingfengbo.top/my-library
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthorCreation:
      type: object
      required: [name, name_cn, nation]
      properties:
        name: { type: string }
        name_cn: { type: string }
        nation: { type: string }
        dynasty: { type: string }
        intro: { type: string }
        photo: { type: string }
    AuthorUpdate:
      type: object
      properties:
        name: { type: string }
        name_cn: { type: string }
        nation: { type: string }
        dynasty: { type: string }
        intro: { type: string }
        photo: { type: string }
    AuthorResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        name_cn: { type: string }
        nation: { type: string }
        dynasty: { type: string, nullable: true }
        intro: { type: string, nullable: true }
        photo: { type: string, nullable: true }
        books:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/BookSimple"

    # ── Book Simple (used in collections/authors) ──
    BookSimple:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        title_cn: { type: string, nullable: true }
        thumb_image: { type: string, nullable: true }
        isbn: { type: string, nullable: true }
        authors:
          type: array
          nullable: true
          items: { type: string }

    # ── Publisher ──
...
paths:
...
  /api/authors:
    get:
      tags: [Authors]
      summary: List authors (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: sort_by, in: query, schema: { type: string, default: name, enum: [name, created_at] } }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  authors:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuthorResponse"
                  total_pages: { type: integer }
                  total_authors: { type: integer }
    post:
      tags: [Authors]
      summary: Create an author
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorCreation"
      responses:
        "200":
          description: Created author
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorResponse"
  /api/authors/{author_id}:
    get:
      tags: [Authors]
      summary: Get author details (with books)
      parameters:
        - { name: author_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Author with books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorResponse"
    put:
      tags: [Authors]
      summary: Update an author
      security:
        - BearerAuth: []
      parameters:
        - { name: author_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorUpdate"
      responses:
        "200":
          description: Updated author
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorResponse"
    delete:
      tags: [Authors]
      summary: Delete an author
      security:
        - BearerAuth: []
      parameters:
        - { name: author_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

  # ═══════════════════════════════════════════════
  # Publishers
  # ═══════════════════════════════════════════════
...